requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

motor==3.4.0
beanie==1.25.0

pydantic==2.7.1
pydantic-settings

torch==2.3.0
transformers==4.41.1
diffusers==0.27.2
accelerate==0.30.1
safetensors==0.4.3
